Saving work.

+++ lib/python/plasma.py
@@ -88,7 +88,7 @@ class PlasmaClient(object):
     self.client.plasma_contains.restype = None
     self.client.plasma_seal.restype = None
     self.client.plasma_delete.restype = None
-    self.client.plasma_evict.restype = ctypes.c_int
+    self.client.plasma_evict.restype = ctypes.c_int64
     self.client.plasma_subscribe.restype = ctypes.c_int
 
     self.buffer_from_memory = ctypes.pythonapi.PyBuffer_FromMemory
